getting the app to run rectify vd
I had trouble getting rectify vd to run in the app. My solution was to call rectify vd as an external fuction. I removed the getOCMparameters function and got it to use the variables in define camera parameters. I put a keyboard in rectify vd as that was my last stopping point before the upload. It now recognizes the variables info, icp, and beta0 from the file.